split job and results loading functions into system
