+++ internal/system/job.go
@@ -1,0 +1,59 @@
+package system
+
+import (
+	"fmt"
+	"strings"
+
+	"github.com/filecoin-project/bacalhau/internal/types"
+)
+
+type ResultsList struct {
+	Node   string
+	Cid    string
+	Folder string
+}
+
+func GetJobData(host string, port int, jobId string) (*types.Job, error) {
+	args := &types.ListArgs{}
+	result := &types.ListResponse{}
+	err := JsonRpcMethod(host, port, "List", args, result)
+	if err != nil {
+		return nil, err
+	}
+
+	for _, jobData := range result.Jobs {
+		if strings.HasPrefix(jobData.Id, jobId) {
+			return jobData, nil
+		}
+	}
+
+	return nil, fmt.Errorf("Could not find job: %s", jobId)
+}
+
+func GetJobResults(host string, port int, jobId string) (*[]ResultsList, error) {
+
+	job, err := GetJobData(host, port, jobId)
+
+	if err != nil {
+		return nil, err
+	}
+
+	results := []ResultsList{}
+
+	for node := range job.State {
+
+		cid := ""
+
+		if len(job.State[node].Outputs) > 0 {
+			cid = job.State[node].Outputs[0].Cid
+		}
+
+		results = append(results, ResultsList{
+			Node:   node,
+			Cid:    cid,
+			Folder: GetResultsDirectory(job.Id, node),
+		})
+	}
+
+	return &results, nil
+}
